more appengine goodness
