+++ glide.yaml
@@ -1,5 +1,8 @@
 package: github.com/lestrrat/sharaq
 import:
+- package: cloud.google.com/go
+  subpackages:
+  - storage
 - package: github.com/bradfitz/gomemcache
   subpackages:
   - memcache
@@ -15,12 +18,18 @@ import:
   subpackages:
   - listener
 - package: github.com/pkg/errors
+- package: golang.org/x/oauth2
+  subpackages:
+  - google
+- package: golang.org/x/sync
+  subpackages:
+  - errgroup
+- package: google.golang.org/api
+  subpackages:
+  - option
+- package: google.golang.org/appengine
+  subpackages:
+  - memcache
 - package: gopkg.in/go-redis/cache.v5
 - package: gopkg.in/redis.v5
 - package: gopkg.in/vmihailenco/msgpack.v2
-- package: github.com/stretchr/testify
-  subpackages:
-  - assert
-- package: cloud.google.com/go
-  subpackages:
-  - storage
