+++ interface.go
@@ -1,11 +1,15 @@
 package sharaq
 
 import (
-	"context"
 	"net/http"
 	"net/url"
 	"regexp"
+	"time"
 
+	"github.com/lestrrat/sharaq/aws"
+	"github.com/lestrrat/sharaq/fs"
+	"github.com/lestrrat/sharaq/gcp"
+	"github.com/lestrrat/sharaq/internal/context"
 	"github.com/lestrrat/sharaq/internal/transformer"
 	"github.com/lestrrat/sharaq/internal/urlcache"
 )
@@ -14,7 +18,11 @@ type Server struct {
 	backend     Backend
 	config      *Config
 	cache       *urlcache.URLCache
+	bucketName  string
+	listenAddr  string
+	logConfig   *LogConfig
 	transformer *transformer.Transformer
+	whitelist   []*regexp.Regexp
 }
 
 type Backend interface {
@@ -23,13 +31,32 @@ type Backend interface {
 	Delete(context.Context, *url.URL) error
 }
 
-// Dispatcher is responsible for marshaling the incoming request
-// to the appropriate backend.
-type Dispatcher struct {
-	cache      *urlcache.URLCache
-	backend    Backend
-	bucketName string
-	listenAddr string
-	logConfig  *LogConfig
-	whitelist  []*regexp.Regexp
+type LogConfig struct {
+	LogFile      string
+	LinkName     string
+	RotationTime time.Duration
+	MaxAge       time.Duration
+	Offset       time.Duration
+}
+
+type DispatcherConfig struct {
+	Listen    string     // listen on this address. default is 0.0.0.0:9090
+	AccessLog *LogConfig // dispatcher log. if nil, logs to stderr
+}
+
+type BackendConfig struct {
+	Amazon     aws.Config // AWS specific config
+	Type       string     // "aws" or "gcp" ("fs" for local debugging)
+	FileSystem fs.Config  // File system specific config
+	Google     gcp.Config // Google specific config
+}
+
+type Config struct {
+	filename   string
+	Backend    BackendConfig
+	Debug      bool
+	Dispatcher DispatcherConfig
+	Presets    map[string]string
+	URLCache   *urlcache.Config
+	Whitelist  []string
 }
