+++ internal/util/request_standalone.go
@@ -1,0 +1,13 @@
+// +build !appengine
+
+package util
+
+import (
+	"net/http"
+
+	"github.com/lestrrat/sharaq/internal/context"
+)
+
+func RequestCtx(r *http.Request) context.Context {
+	return r.Context()
+}
